Register and User Login

+++ base/views.py
@@ -12,9 +12,14 @@ from .forms import RoomForm
     {'id':1, 'name': 'Lets learn Python'},
     {'id':2, 'name': 'Design with me'},
     {'id':3, 'name': 'Frontend Developers'},
+    ]
 '''
 
 def loginPage(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
 
@@ -35,13 +40,18 @@ def loginPage(request):
         else:
             messages.error(request, 'Invalid username or password.')
 
-    context={}
+    context={'page': page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
 
+def registerPage(request):
+    page = 'register'
+
+    context={'page': page}
+    return render(request, 'base/login_register.html')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -83,7 +93,7 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
-    if request.user != room.user:
+    if request.user != room.host:
         return HttpResponse("You ain't the owner though GTFO!")
 
     if request.method == 'POST':
@@ -99,6 +109,10 @@ def updateRoom(request, pk):
 @login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You ain't the owner though GTFO!")
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
